requirements.txt
PyQt6
python-pptx
markdown
markitdown
spire.presentation
